[ticket/10844] Remove phpBB root path in sanitise when searching for styles
Remove phpBB root path from the beginning of the directory name, when we
try to find styles. Styles have the root path prepended when they are
added to the locator. If we do not remove it here, template files from
extensions can not be used, when the current root path is not the phpBB
root path.

PHPBB3-10844