[ticket/11465] Add unit tests for acp_modules::get_module_infos()
The tests add 3 different modules. One acp module that should be found
(acp/a_module), one acp module that should not be found (acp/fail_module),
and one mcp module that should work again (mcp/a_module). The modules'
info files had to be included as they were not auto-loaded for some
reason.
There are several test stages. First of, it is tested if the correct mcp
and acp module is returned. Afterwards, the proper loading of specified
modules is tested. One with an existing module and one with a not existing
module. Finally, the test concludes with trying to get the module info of
not existing ucp modules. Other classes like foobar would have also worked
for that check but I decided to use the ucp type of class as that is the
one type missing from the added test modules.

PHPBB3-11465