[ticket/9975] Moved a few E_USER_ERROR errors to /language
There were a few error messages that a user could experience that would, previously, be without any the ability to be localized. There are some more E_USER_ERRORs that I did not change to a constant, for example the error message that is displayed if there aren't any folders in /language.

PHPBB3-9975