[ticket/11129] Added headers to empty subscription state
"You are not subscribed to any forums".
While correct, could lead to confusion. What's missing is a header right above the message that says "Watched Topics" to match the "Watched Forums" header above. This will assert the idea that there are two separate watchable sections.
As is, it looks like that message is talking about the "Watched Forums" section, since it's listed under that header.

PHPBB3-11129