[ticket/11531] Add basic set of functional tests for new avatar system
This currently includes tests for the gravatar, remote avatar, and upload
avatar drivers. The local avatar driver is currently not supported. Two
tests currently fail and have been marked as incomplete. One failure is
due to an issue with Symfony's DomCrawler that exists in Symfony < 2.2.
The other one is caused by a bug in the way remote avatars are treated.
According to my tests, this bug also exists in phpBB 3.0.11 and has been
reported in the tracker: http://tracker.phpbb.com/browse/PHPBB3-11534

PHPBB3-11531