[ticket/11531] Add acp and ucp groups functional tests for avatars
The test for the acp and ucp groups settings are currently marked as
incomplete due to a bug that causes the settings to not show an error
when incorrect data is entered. However, the avatar data is not saved.
That means that even though the error treatment seems to correctly work,
the user is never informed of the issues with the submitted avatar data.

PHPBB3-11531