[ticket/11577] Correct over-sized Topic Rows
Fixes a change made in PR 1331 that oversized topic rows.
Icons in topic rows, in Prosilver, are only 27px high, so we do not need
a min-height as high as 40px. Setting it to 30px is more conservative
and reclaims the original height of topic rows from before PR 1331.
http://tracker.phpbb.com/browse/PHPBB3-11577

PHPBB3-11577