[feature/twig] Fixing template events test (and behavior)
According to the test, template event behavior was never correct.

Only ONE template event file is supposed to be included from EACH
extension. As it was before, EVERY matching template event file from
each extension was included (this was how it was designed).

E.g.
Event call in prosilver "foo"
Extension has template "foo" in prosilver AND all

foo from all would be included, then foo from prosilver would be included

This was clearly not designed correctly as only the most specific event
file from each extension should be loaded, otherwise events could only
ever be put in a single style tree (either only all, or only prosilver
and subsilver2 and any style that inherits from neither of those).
Otherwise the events would be duplicated on output (which is clearly not
desirable).

The Twig behavior already was correct as I designed it, so only the one
most specific template event file found would be included from each
extension. The tests had to be updated for the correct expected output.

PHPBB3-11598