[feature/api] Added forum entities and models
In this change I added entities to store all fields of
a forum row as an object, later to be serialized using
Symfonys serializer.

A model for API/forum was created to handle data retrieval
of forums.

Controller was split up to be a controller per API, this time
a controller for the API regarding forums.

Services for nestedtree_forums were added as well as for the
API Forum model.

PHPBB3-11595