[feature/api] Implemented the Symfony serializer
A normalizer for the forum entity has been created
with the function of turning the forum entity
object into an array using only the relevant fields
which can then be serialized into either json
or xml (only json for now). This has been done by
using Symfonys Serializer component.

The entity class now uses a getter and a setter.

PHPBB3-11595