[feature/twig] Fix template/template_php_test.php
Must create a template file in the cache to load as this causes errors
otherwise. The problem was that Twig builds template files into classes,
which are always stored in PHP memory after being loaded. Because of this,
Twig would never recompile a template that was already compiled on the same
page load (so switching enable PHP on/off in two tests would not work).

PHPBB3-11598