[ticket/11990] Remove result_mssqlnative usage in acp_database.php
The class result_mssqlnative was removed in ticket 11980. This removes
the dependency on that class from the database backup utility. It is
possible to use the sqlsrv functions to retrieve the field information,
but they must be then mapped back to type names. That mapping was removed
in 11980 and it is easier to just look it up in the information schema
table.

PHPBB3-11990