[ŧicket/11896] Set form_time with time() when marking all notifications read
form_time is only set if is passed via the form. Since the mark notifications
read link does not use the form, it will default to 0 causing the mark
notifications logic to only mark notifications read if their time is smaller
or equal to 0. This patch will change ucp_notifications to correctly set
form_time for the confirm_box.

PHPBB3-11896