[ticket/11997] Use functional test cases that should always work
The previous test cases that tried to redirect to ../index.php and similar
might cause us to try to go to http://localhost/../index.php, which will
result in http://index.php. As this obviously will trigger an error as
intended, we should not put this inside our test cases for the redirect
function.

PHPBB3-11997