[ticket/9040] Count HTML entities as single in custom profile fields
Currently when an user adds a HTML entity to a custom profile field, the length
gets incremented by 4 since the code reads &gt; instead of <. However, the
length is presentational and not DB storage so it should be treated as a
single character even if it takes four lengths in the DB by being stored as
a HTML entity.

Work around this by decoding html entities before counting. Also, added unit
tests for string field type

PHPBB3-9040