[ticket/12303] Add "rules-link" class to rules-link
When displaying forum rules, there is no way in proSilver to distinguish
between a forum-rules "link", and regular forum rules text.

This makes it very difficult to style the rules-link, because it will
create a conflict with any <a> element inside regular forum rules text
(which get the .postlink class).