[ticket/12270] Correct post&topic approval msg, no template changes
Following the previous commits, this patch removes the changes
done to templates to depend only on code, as suggested by
@nickvergessen.
All previous modifications have been removed, with another one
which checks, in the approve_posts() function, if it is sending a
notification for post approval and it is the first one in the
topic, send notification indicating that the topic has been
approved (and not the post). If not the first post in the topic,
do just as before (send a notification about a newly approved
post).

PHPBB3-12270