[ticket/12361] Replace the Google logo.
Currently, a link to Google and its logo are displayed in the
language/en/help_bbcode.php file to explain how to correctly format tags
and to add an image to a post. However, the Google logo is outdated and
Google does not allow the use of its logo on third-party websites. Plus,
that overload their servers. I used the blue phpBB logo to replace it
because the grey one with transparent background is hardly visible.

Originally, I just refreshed the old Google logo with the new one
(https://www.google.com/images/srpr/logo5w.png) but Joas proposed to use
the phpBB logo, and I agree with him.

PHPBB3-12361