[ticket/12393] Open quickedit window whenever edit is clicked on post
This will be another slight change in behavior. When clicking the edit button
on a post the quickedit window will appear. Clicking the edit button of a
different post will cancel the initial quickedit and open another quickedit
window in the post the edit button was clicked in instead of going to the
full editor as introduced by the previous commit.

PHPBB3-12393