[ticket/11366] Extension's version's check on details
Add a feature to check automatically the version of the installed
extensions. The informations are cached for 24 hours (like for the
global update check on the main page of the acp).

The informations about the versions are display both on the global list
and on the detailled page.

To do this the developper has to to let the composer.json of the latest
version available and add some informations into it :

    "extra": {
		"version-check": {
			"host": "<the host>",
			"directory": "<the directory containing the file>",
			"filename": "<the composer.json of the latest version>"
		}
    }

He can also add two extra informations which will be displayed if a new
version is available :

	"extra": {
		"download": "<download link>",
		"annoucement": "<announcement link>",
    }

Currently a notice is displayed when the "extra.version-check"
informations are
missing.

Ticket: https://tracker.phpbb.com/browse/PHPBB3-11366

Signed-off-by: Nicofuma <github@nicofuma.fr>

PHPBB3-11366