[ticket/12501] Only select attachments uploaded before the post was reported.
The post text displayed to the user is the same as when it was reported, even if
the post was edited afterwards, thus it follows that the attachments would be
the same. This also addresses an issue where the index in the attachment bbcode
no longer matches that of its corresponding attachment when new attachments
are uploaded after the post is reported.

PHPBB3-12501