[ticket/12673] PLUPLOAD fails due missing constant IMAGETYPE_SWC
A PHP Warning is issued when you try to uplaod Image based file with plupload.
[phpBB Debug] PHP Notice: in file <b>[ROOT]/includes/functions_upload.php
on line1067: Use of undefined constant IMAGETYPE_SWC - assumed 'IMAGETYPE_SWC'
It changes the whay JSON is parssed if the PHP NOtices are unabled.

A little change to image_types() function checking if global constants are
defined and defining if some are not.
As IMAGETYPE_* are known values.

PHPBB3-12673