[ticket/12786] Extend profilefield_base_migration.php class
If extensions authors will use profilefield_base_migration.php
class as base class for creation of CPFs, the class need to be
extended a bit.

At the moment there are two issues with it:

It does not create entries for PROFILE_FIELDS_LANG_TABLE
(as it can't know what fields to create).

The migration is not "cleaning" the DBMS after purging
the extension and leaves everything except the CPF column.

Adding profilefield_language_data

PHPBB3-12786