[ticket/11768] Invalidate the text_formatter cache whenever a style is installed
The acp_styles module purges the cache every time a style is uninstalled,
modified or made default, but it does not purge the cache when a new style
is installed. Here we invalidate the text_formatter cache (not purge the
whole cache) so that new styles take effect immediately.

PHPBB3-11768