[feature/SCEditor] Adding xsl_parse_helper; Basic XSL parse to get HTML
Extracts the HTML out of the XSL and locates all XSL-defined variables.
Generates a tree built using php arrays with all relevant data and metadata
of the parsed information on BBCode definitions and translations.

Unfortunately, it is API-coupled with TextFormatter

This class is supposed to be agnostic to any WYSIWYG editor that is used.

PHPBB3-10418