[ticket/13867] Enable/disable mechanism for new profile field types
Adds methods to enable, disable and purge profile field types to the
profilefields\manager class.  These methods are to be called from
extensions that add new profile field types on the enable, disable
and purge methods of the ext class.  If not done, any profile field
of a new type that is left after extension is disabled or removed will
break the forum in several places. Typo fixed.

PHPBB3-13867