[feature/SCEditor] Add override options and control
These overrides and code is, for now, meant to be used only by extensions
and for internal code.
Later on, after testing is made, it can change to also include ACP options,
either on phpBB 3.2 or phpBB 3.3.
Regardless of that, I don't expect the interface to have them until this is
merged

PHPBB3-10418