[feature/SCEditor] Quote BBCode definition changes
Due to the WYSIWYG editor requirements, the quote tag has to be slightly
different. Same will happen to the code tag.
I don't know how to do this right in a different way.
If you have a good idea, please step up.

As for the "data-sce-*" stuff, those are meant to be replaced with
"data-*" and the sce class identifies them and translates or they may be
removed altogether and calculated automatically.
Let's see how it goes.

PHPBB3-10418