[ticket/14395] Add event core.viewtopic_add_quickmod_option_after
Added an event before the function phpbb_add_quickmod_option is called in
the viewtopic.php file. The function is run before any event can be called to
modify ... meaning that any modifications done to this array
will not affect the options displayed to the user on the viewtopic page itself.
This event serves two purposes - one, it is easier to add new quickmod options,
and two, it allows an extenstion developer to check additional permissions and
apply further authentication for the display of the quickmod options.

PHPBB3-14395