[ticket/14395] core.viewtopic_add_quickmod_option_before
Added an event before the function phpbb_add_quickmod_option is called in
the viewtopic.php file. This event serves two purposes - one, it is easier to
add new quickmod options, and two, it allows an extenstion developer to check
additional permissions and apply further authentication for the display of the
quickmod options.

PHPBB3-14395