[ticket/14581] Add core events to content_visibility.
Events added for both changing the content visibility for both posts and topics,
executing right before and after the SQL UPDATE query for setting the
visibillity. Changed the _after events to be after all of the sync functions,
updated the comments to be mre accurate, and removed the _sql from the _after
events to more accurately reflect their new position.

PHPBB3-14581