[audio-support] Add support for playing audio files directly in phpBB
I've tested this with MP3 and M4A files with Firefox, Chrome and Safari.
Everything works as expected. Extending the change to support video
tags is trivial, but I left it out as I had no need for it.

Safari on OS X is quite picky about which files to play, and refuses
to play MP4 audio unless the mime type is set on the request.

The one thing missing is migration.

PHPBB3-14771