[ticket/14733] Make sure detect_algorithm() works correctly and add tests
detect_algorithm() returned array() if an algorithm prefix was more than 2
characters long. This might have been invalid for other prefixes. In order
to correctly cope with other prefixes, another check for a backslash in the
prefix definitino has been added.
This was discovered while writing the tests for the newly added interface.

PHPBB3-14733