[ticket/13853] Flexible schema for profilefields step 1 configuration
Modify the way adm/style/acp_profile.html STEP 1 (lines 106 to 128)
is coded to make it flexible and expandable to other new types of
profile fields, rather than the current mechanism, hardcoded for
the default types.

A flexible mechanism similar to the one used of STEP 2 is used.

This required a slight modification of the type_interface, and
corresponding modifications to all implementations of that interface.

Fixed an unrelated bug in type_interface, get_field_name was missing.

PHPBB3-13853