[ticket/15050] Use new file when new file already exists
Instead of trying to diff the new file against a pre-existing file,
we're just going to use the new file. It's impossible to know whether
the pre-existing file is newer or older than the new file. As the
system will rely on the files being in the "new" state it's better
to simply use the file in "new" state.

PHPBB3-15050