[ticket/15011] All errors on metadata throw exceptions
There is inconsistency in the way error are treated in metadata_manager.
Some method return false on error, others throw exception.  With this,
the usage is homogeneus, and the uses of these are also adapted.
Using same ticket as it is the same issue, solved in a different way.

PHPBB3-15011