[PHPBB3-15133] filespec.php getimagesize returns 0
in some cases jpeg files can cause the builtin php getimagesize to return
a width or height of 0.  In these cases '!isset' is a more correct validation
because it doesn't fail where no value returned is still caught.

'empty' counts 0 as 'empty' even though there IS a value.  Potentially replacing
the part of filespec that needs to use getimagesize would be better, but
this is a specifically useful fix for the situation.