[ticket/14498] Don't display 'Who is online' members when permission is No
When the user permission 'Can view profiles, memberlist and online list' is
set to No (or Never) the user should not be able to see which members are
online in the following places, index page, view topic and view forum.

Before this change, guests and bots would see the list of members who are
online, which doesn't match with the user permission and could create privacy
issues where guests or bots could track when a member was online.

PHPBB3-14498