[ticket/15397] Check that table exist before modifying its columns/indexes/keys
Given a migration, A, that adds database tables and drops them when reverting,
and a migration, B, that modifies the columns of a table added by migration A,
the migrator runs in the correct order when purging the extension but drops the
first table created by migration A, the starts over with migration B, which
then returns an error because the table no longer exists. This ensures that the
table exists before trying to modify its columns.

PHPBB3-15397