SMTP server address AND PROTOCOL
Many people reading "SMTP server address", think they know what to do and put "smtp.(...).com" automatically, not reading the explanation (because there are too many explanations which do not add new meanings, and maybe people get tired or it's hard for them to read). So adding this vital part to the title improves accessibility and saves hours of struggling with errors.