[ticket/15649] Allow to define a custom cache directory
(Experimented) users should be allowed to easily change the
location of the cache directory.

Some examples of use:
* Move the cache directory to a ramdisk filesystem, allowing to
reduce writes on drives
* On development boards using different config.php files (to easily
switch from a "profile" to another one), it allows to use a
different cache directory for each profile

This commit allows to define a constant PHPBB_CACHE_PATH
(ideally, in config.php) which will be used to define the
path of the cache directory.

PHPBB3-15649