[ticket/15683] Better error message when commit message has CRLF
When someone tries to do a pull request with incorrect line breaks in the
commit message, the error message as given by Travis is "Unexpected EOF
encountered". The error message should be clearer that CR/LF was the
problem. This change adds a separate check for CRLF line endings.

PHPBB3-15683