[ticket/15756] Add Lost Username function
This patch adds to the login page and (logged-out) index page a
link 'I forgot my username'.  This link brings up a form for
retrieving username(s) by entering the registered email address.
The response(s) is sent by email, or jabber, as configured for the
user.
For boards that allow multiple accounts per email address, a
username reminder is sent for each account sharing the address.

PHPBB3-15756