[ticket/15257] Provide error messages for non enableable extensions
A function is added to the extension interface to be able to retrieve
the error messages generated in the extension.

The core extension error message is always put in the beginning of the
error array, for clarity. phpBB's error message goes first —which is
internationalized and easy to understand— and then the developer's
custom error messages.