[ticket/16193] Re-implement drafts
See https://www.phpbb.com/community/viewtopic.php?f=436&t=2522421

This improvement will allow draft posts and PMs to be saved with attachments and formatting.

It will remove the drafts table and about 95% of the existing code related to drafts.
The code added to replace (and enhance) the behaviour should be somewhat more compact.

PHPBB3-16193