[ticket/16363] Allow replying to locked topics if user has m_lock permissions
An user may not be able to edit - but lock and unlock.
It is terribly inconvenient for users being only able to un/lock threads to
have to first unlock a topic, then reply and then re-lock it.

Thus in reply mode we check for lock permissions instead of edit.

This also includes editing if the user could normally edit his post if it
weren't for the lock.

PHPBB3-16363