[ticket/16722] Avoid duplicate "app.php" in generated URLs
When enable_mod_rewrite = false and force_server_vars = false, the base URL is
taken from getBaseUrl, which as specified in this method's documentation keeps
the script filename at the end of the URL.
This means that the URL contains "/app.php" at the end, which is not expected
by the following code, and results in "app.php/app.php" in the generated URL.

When force_server_vars = true, the base URL is taken from the script_path
config, which contains only the folder of the phpBB install, and the bug does
not appear.

This commit aligns the behaviour when force_server_vars = false, by calling
Symfony's getBasePath instead of getBaseUrl.

PHPBB3-16722