[ticket/8777] Discourage deleting user with only one group
It will still be possible from the manage groups page though is discouraged
now by hiding the delete entry when managing users.

In addition to this, the manage user page will now show that no group is
specified for the user when viewing the groups page of a user.

PHPBB3-8777