[ticket/17365] Enforce the search word limit on queries containing operators without whitespace
Provided `$countable_keywords` wherein the existing `$keywords` value is modified so that any `-`, `+` and `|` characters without preceding spaces is replaced with the same but with a space in front of each. 

These spaces allow the string to be more accurately split when used instead of $keywords inside the $num_keywords calculation. 

This prevents the word limit being bypassed in search by the use of operators without whitespace.

PHPBB-17365